:snail: 小修改

+++ lib/src/search.rs
@@ -152,7 +152,7 @@ impl<'a, R: Rule> World<'a, R> {
     ///
     /// The difference between `step` and `self.steps` is that the former
     /// will be resetted in each `search`.
-    fn go(&mut self, step: &mut usize) -> bool {
+    fn go(&mut self, step: &mut u64) -> bool {
         loop {
             *step += 1;
             if self.proceed() {
@@ -193,7 +193,7 @@ impl<'a, R: Rule> World<'a, R> {
     /// `None` if such pattern does not exist,
     /// `Searching` if the number of steps exceeds `max_step`
     /// and no results are found.
-    pub fn search(&mut self, max_step: Option<usize>) -> Status {
+    pub fn search(&mut self, max_step: Option<u64>) -> Status {
         let mut step_count = 0;
         if self.get_unknown(0).is_none() && !self.backup() {
             return Status::None;
@@ -229,7 +229,7 @@ pub trait Search {
     /// `None` if such pattern does not exist,
     /// `Searching` if the number of steps exceeds `max_step`
     /// and no results are found.
-    fn search(&mut self, max_step: Option<usize>) -> Status;
+    fn search(&mut self, max_step: Option<u64>) -> Status;
 
     /// Displays the whole world in some generation.
     ///
@@ -245,7 +245,7 @@ pub trait Search {
     fn gen0_cell_count(&self) -> usize;
 
     /// Number of conflicts during the search.
-    fn conflicts(&self) -> usize;
+    fn conflicts(&self) -> u64;
 
     /// Set the max cell counts.
     ///
@@ -256,7 +256,7 @@ pub trait Search {
 
 /// The `Search` trait is implemented for every `World`.
 impl<'a, R: Rule> Search for World<'a, R> {
-    fn search(&mut self, max_step: Option<usize>) -> Status {
+    fn search(&mut self, max_step: Option<u64>) -> Status {
         self.search(max_step)
     }
 
@@ -272,7 +272,7 @@ impl<'a, R: Rule> Search for World<'a, R> {
         self.gen0_cell_count
     }
 
-    fn conflicts(&self) -> usize {
+    fn conflicts(&self) -> u64 {
         self.conflicts
     }
