+++ web/src/worker.rs
@@ -7,7 +7,7 @@ use yew::{
     Callback,
 };
 
-const VIEW_FREQ: usize = 50000;
+const VIEW_FREQ: u64 = 50000;
 
 struct Job {
     timeout: TimeoutService,
